Fix countless warnings with the test suite (#64)
* Fix `@error_handler` not initialized warning

```
................./home/travis/build/Rykian/clockwork/lib/clockwork/manager.rb:40: warning: instance variable @error_handler not initialized
304......../home/travis/build/Rykian/clockwork/lib/clockwork/manager.rb:40: warning: instance variable @error_handler not initialized
```

* Fix mocha require warning

We just have to change to using mocha/minitest here per the warning and their documentation.

```
Mocha deprecation warning at /home/travis/build/Rykian/clockwork/test/at_test.rb:3:in `require': Require 'mocha/test_unit', 'mocha/minitest' or 'mocha/api' instead of 'mocha/setup'.
```

* Fix interpreter warning on argument prefix

This resolves the following:
```
/home/travis/build/Rykian/clockwork/test/database_events/support/active_record_fake.rb:32: warning: `*' interpreted as argument prefix
```

* Fix countless `@name` not initialized warnings

The test suite was displaying tons of warnings in regards to this name accessor. There were 2 different kinds.

The first was the most frequent:
```
/home/travis/build/Rykian/clockwork/test/database_events/test_helpers.rb:42: warning: instance variable @name not initialized
```

And the other was for declaring the getter method override which results in 3 warnings.

```
/home/travis/build/Rykian/clockwork/test/database_events/test_helpers.rb:41: warning: method redefined; discarding old name
276/home/travis/build/Rykian/clockwork/test/database_events/test_helpers.rb:50: warning: method redefined; discarding old name
277/home/travis/build/Rykian/clockwork/test/database_events/test_helpers.rb:64: warning: method redefined; discarding old name
```

While it leads to be a bit more code this seemed the safest and easiest way to make the change with the test helpers and not have to update existing tests and also remove all noisy warnings.